playground/requirements.txt
elasticsearch==8.11.0
Faker==20.1.0
python-dotenv==1.0.0
fire==0.5.0